# docs/swagger.yaml
basePath: /
definitions:
...
  review.ReviewParams:
    properties:
      comment:
        type: string
      rating:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Akshay Sharma
    url: github.com/akshay4git/Go-Eats
  description: A food delivery backend system with user management, restaurant operations,
    cart/order processing, delivery management with 2FA and real-time notifications.
  title: GO-EATS API
  version: "1.0"
paths:
...
  /restaurant/:
    get:
      description: Returns a list of all restaurants
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all restaurants
      tags:
      - Restaurant
    post:
      consumes:
      - multipart/form-data
      description: Creates a new restaurant with photo upload (multipart/form-data)
      parameters:
      - description: Restaurant photo
        in: formData
        name: file
        required: true
        type: file
      - description: Restaurant name
        in: formData
        name: name
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: City
        in: formData
        name: city
        required: true
        type: string
      - description: State
        in: formData
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a restaurant
      tags:
      - Restaurant
  /restaurant/{id}:
    delete:
      description: Deletes a restaurant by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a restaurant
      tags:
      - Restaurant
    get:
      description: Returns a single restaurant by its ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get restaurant by ID
      tags:
      - Restaurant
...
  /review/{restaurant_id}:
    get:
      description: Returns all reviews for a specific restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Adds a review for a restaurant by the authenticated user
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Review details
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/review.ReviewParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a review
      tags:
      - Reviews
  /review/{review_id}:
    delete:
      description: Deletes a review by ID for the authenticated user
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - Reviews
  /user/:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - User
  /user/{id}:
    delete:
      description: Deletes a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"